requirements.txt
aenum                  3.1.15
aiohttp                3.9.1
aiosignal              1.3.1
allure-pytest          2.13.2
allure-python-commons  2.13.2
attrs                  23.2.0
certifi                2023.11.17
charset-normalizer     3.3.2
dill                   0.3.7
et-xmlfile             1.1.0
execnet                2.0.2
Faker                  22.2.0
frozenlist             1.4.1
h11                    0.14.0
idna                   3.6
iniconfig              2.0.0
Jinja2                 3.1.3
MarkupSafe             2.1.3
multidict              6.0.4
mysql-connector-python 8.2.0
openpyxl               3.1.2
outcome                1.3.0.post0
packaging              23.2
pip                    23.2.1
pluggy                 1.3.0
protobuf               4.21.12
PySocks                1.7.1
pytest                 7.4.4
pytest-html            4.1.1
pytest-metadata        3.0.0
pytest-reportportal    5.3.1
pytest-xdist           3.5.0
python-dateutil        2.8.2
PyYAML                 6.0.1
reportportal-client    5.5.4
requests               2.31.0
selenium               4.16.0
selenium-page-factory  2.6
six                    1.16.0
sniffio                1.3.0
sortedcontainers       2.4.0
trio                   0.24.0
trio-websocket         0.11.1
urllib3                2.1.0
wsproto                1.2.0
yarl                   1.9.4
